backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
yt-dlp==2024.3.10
python-multipart==0.0.9
pydantic==2.5.3
python-dotenv==1.0.0
certifi>=2023.7.22
urllib3>=2.0.0
requests>=2.32.0
